pyproject.toml
[tool.poetry]
name = "certbot-dnspod"
version = "0.1.1"
description = "A certbot plugin for DNSPod"
authors = ["codeif <me@codeif.com>"]
license = "BSD-3-Clause"
readme = "README.md"
repository = "https://github.com/codeif/certbot-dnspod"

[tool.poetry.dependencies]
python = "^3.8"
certbot = "^1.23.0"
"zope.interface" = "^5.4.0"
dnspod-sdk = "^0.0.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
black = "^22.1.0"

[tool.black]
line-length = 120
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'

[tool.poetry.plugins."certbot.plugins"]
"certbot-dnspod" = "certbot_dnspod:Authenticator"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
